requirements.txt
polars>=0.20.0
requests>=2.31.0
feedparser>=6.0.11
beautifulsoup4>=4.12.2
lxml>=5.1.0
trafilatura>=1.8.0
tenacity>=8.2.3
python-dateutil>=2.9.0
pyyaml>=6.0.1
